+++ common/src/main/java/com/lne_paladins/effect/LNE_PaladinsEffects.java
@@ -1,7 +1,10 @@
 package com.lne_paladins.effect;
 
+import net.minecraft.entity.attribute.EntityAttributeModifier;
+import net.minecraft.entity.attribute.EntityAttributes;
 import net.minecraft.entity.effect.StatusEffectCategory;
 import net.minecraft.util.Identifier;
+import net.spell_engine.api.config.AttributeModifier;
 import net.spell_engine.api.config.ConfigFile;
 import net.spell_engine.api.config.EffectConfig;
 import net.spell_engine.api.effect.*;
@@ -27,21 +30,26 @@ public class LNE_PaladinsEffects {
                     )
             )
     ));
-    public static Effects.Entry HOLY_WEAPON = add(new Effects.Entry(Identifier.of(MOD_ID, "holy_weapon"),
-            "Holy Weapon",
-            "Enhances your melee hits with heals and additional healing power damage, 100% crit chance against undead.",
-            new CustomStatusEffect(StatusEffectCategory.BENEFICIAL, 0xffffcc),
-            new EffectConfig(
-                    List.of(
-                    )
-            )
-    ));
     public static Effects.Entry PREVENTION = add(new Effects.Entry(Identifier.of(MOD_ID, "holy_prevention"),
             "Prevention",
             "Heals you when you fall below 20% max health, the heal amount enhances with effect amplifier. If you´re already under 20% when the effect is applied, the healing is reduced.",
             new PreventionStatusEffect(StatusEffectCategory.BENEFICIAL, 0xffffcc),
             new EffectConfig(
                     List.of(
+                    )
+            )
+    ));
+    public static Effects.Entry TEMPLARS_RETRIBUTION = add(new Effects.Entry(Identifier.of(MOD_ID, "templars_retribution"),
+            "Templars Retribution",
+            "Increases melee damage.",
+            new CustomStatusEffect(StatusEffectCategory.BENEFICIAL, 0xffd700),
+            new EffectConfig(
+                    List.of(
+                            new AttributeModifier(
+                                    EntityAttributes.GENERIC_ATTACK_DAMAGE.getIdAsString(),
+                                    0.2F,
+                                    EntityAttributeModifier.Operation.ADD_MULTIPLIED_BASE
+                            )
                     )
             )
     ));